+++ NodeJs/Node_Module/myModule.js
@@ -1,0 +1,13 @@
+let count = 0;
+
+const inc = () => ++count;
+const dec = () => --count;
+
+const getCount = () => count;
+
+
+module.exports = {
+    inc,
+    dec,
+    getCount
+}
